Removed debugging statements
