+++ index.js
@@ -8,6 +8,7 @@ const unixify = require('unixify')
 const mutexify = require('mutexify')
 const duplexify = require('duplexify')
 const through = require('through2')
+const pump = require('pump')
 
 const coreByteStream = require('hypercore-byte-stream')
 const MountableHypertrie = require('mountable-hypertrie')
@@ -279,7 +280,6 @@ class Hyperdrive extends EventEmitter {
     })
 
     function oncontent (st, contentState) {
-      console.log('IN ONCONTENT, content length:', contentState.feed.length, 'byteLength:', contentState.feed.byteLength)
       if (st.mount && st.mount.hypercore) {
         var byteOffset = 0
         var blockOffset = 0
@@ -295,24 +295,6 @@ class Hyperdrive extends EventEmitter {
         feed = contentState.feed
       }
       const byteLength = length !== -1 ? length : (opts.start ? st.size - opts.start : st.size)
-
-      console.log('st:', st)
-      console.log('byteLength:', byteLength, 'blockOffset:', blockOffset, 'byteOffset:', byteOffset, 'blockLength:', blockLength)
-
-      /*
-      if (byteOffset === 104236) {
-        return feed.get(169, (err, contents) => {
-          console.log('CONTENTS AT BLOCK 169:', contents)
-          stream.start({
-            feed,
-            blockOffset,
-            blockLength,
-            byteOffset,
-            byteLength
-          })
-        })
-      }
-      */
 
       stream.start({
         feed,
@@ -598,14 +580,12 @@ class Hyperdrive extends EventEmitter {
     if (typeof opts === 'function') return this.stat(name, null, opts)
     if (!opts) opts = {}
 
-    console.log('STATTING:', name)
     this.lstat(name, opts, (err, stat, trie) => {
       if (err) return cb(err)
       if (!stat) return cb(null, null, trie, name)
       if (stat.linkname) {
         if (path.isAbsolute(stat.linkname)) return this.stat(stat.linkname, opts, cb)
         const relativeStat = path.resolve('/', path.dirname(name), stat.linkname)
-        console.log('RELATIVE STAT:', relativeStat, 'NAME:', name, 'LINKNAME:', stat.linkname)
         return this.stat(relativeStat, opts, cb)
       }
       return cb(null, stat, trie, name)
@@ -641,15 +621,11 @@ class Hyperdrive extends EventEmitter {
       createStatStream(this, this._db, name, { ...opts, recursive }),
       through.obj(({ path: statPath, stat }, enc, cb) => {
         const relativePath = (name === statPath) ? statPath : path.relative(name, statPath)
-        console.log('NAME:', name, 'STATPATH:', statPath, 'RELATIVEPATH:', relativePath)
-        const splitPath = relativePath.split('/')
         if (recursive) return cb(null, relativePath)
-        if (name === '/') return cb(null, splitPath[0])
-        return cb(null, splitPath.length > 1 ? splitPath[1] : splitPath[0])
+        return cb(null, relativePath.split('/')[0])
       })
     )
     return collect(nameStream, (err, entries) => {
-      console.log('COLLECTED ENTRIES:', entries)
       if (err) return cb(err)
       return cb(null, entries)
     })
